Puzzle Solver (clock & water)
This solver is made to solve both the clock and water puzzles. The water
puzzle has you creating empty jugs to help obtain the goal gallon amount
declared by you in the beginning. The clock puzzle calculates the steps
needed to get from the beginning to the ending as declared by you as
well.